Normalize validator powers on Ethereum
I don't have a good understanding of how power is managed on Cosmos
but I suspect the number continues to increase as more tokens are
bonded. This is all fine and well if you're aware of every number
from the staking itnerface but not if you're an Ethereum contract
that's just trying to determine if your vote should pass.

This patch normalizes the voting power amounts such that all
validators have their respective percentage of Cosmos voting power
represented roughly equally on Ethereum with a resolution of
uint32Max I think this is high enough resolution to work with.

This code will have to be carefully reviewed and reconsidered later on