WIP add claims storage
This patch actually stores the Ethereum claims. Serializing interface
objects is onerous so what we do is create a Proto definition for a
struct that implements the interface. In this way when you go to
load the claim you can simply take the GenericClaim struct and turn
it back into the interface object you stored.

This patch does not yet include tools for iteration and generally
doesn't work at all. This is because until key delegation is implemented
it's impossible to finish the detail work around what address type is
expected where.